refactor: 시간 측정 및 관리를 담당하는 OperatingTimeManager를 만들어, Bank 객체에 의존성 주입
